Refactored custom event handler, fixed controller issue for owner deletion
