+++ controllers/telemetry/logpipeline_controller.go
@@ -20,21 +20,18 @@ import (
 	"context"
 	"fmt"
 
+	utils "github.com/kyma-project/telemetry-manager/internal/kubernetes"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/util/workqueue"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
-	"github.com/go-logr/logr"
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/telemetry-manager/internal/reconciler/logpipeline"
 	"github.com/kyma-project/telemetry-manager/internal/secretref"
@@ -67,32 +64,32 @@ func (r *LogPipelineReconciler) SetupWithManager(mgr ctrl.Manager) error {
 		For(&telemetryv1alpha1.LogPipeline{}).
 		Watches(
 			&source.Kind{Type: &appsv1.DaemonSet{}},
-			enqueueRequestForOwnerFuncs(ctrl.Log),
+			utils.EnqueueRequestForOwnerFuncs(ctrl.Log),
 			builder.WithPredicates(setup.DeleteOrUpdate()),
 		).
 		Watches(
 			&source.Kind{Type: &corev1.Service{}},
-			enqueueRequestForOwnerFuncs(ctrl.Log),
+			utils.EnqueueRequestForOwnerFuncs(ctrl.Log),
 			builder.WithPredicates(setup.DeleteOrUpdate()),
 		).
 		Watches(
 			&source.Kind{Type: &corev1.ConfigMap{}},
-			enqueueRequestForOwnerFuncs(ctrl.Log),
+			utils.EnqueueRequestForOwnerFuncs(ctrl.Log),
 			builder.WithPredicates(setup.DeleteOrUpdate()),
 		).
 		Watches(
 			&source.Kind{Type: &corev1.ServiceAccount{}},
-			enqueueRequestForOwnerFuncs(ctrl.Log),
+			utils.EnqueueRequestForOwnerFuncs(ctrl.Log),
 			builder.WithPredicates(setup.DeleteOrUpdate()),
 		).
 		Watches(
 			&source.Kind{Type: &rbacv1.ClusterRole{}},
-			enqueueRequestForOwnerFuncs(ctrl.Log),
+			utils.EnqueueRequestForOwnerFuncs(ctrl.Log),
 			builder.WithPredicates(setup.DeleteOrUpdate()),
 		).
 		Watches(
 			&source.Kind{Type: &rbacv1.ClusterRoleBinding{}},
-			enqueueRequestForOwnerFuncs(ctrl.Log),
+			utils.EnqueueRequestForOwnerFuncs(ctrl.Log),
 			builder.WithPredicates(setup.DeleteOrUpdate()),
 		).
 		Watches(
@@ -101,68 +98,6 @@ func (r *LogPipelineReconciler) SetupWithManager(mgr ctrl.Manager) error {
 			builder.WithPredicates(setup.CreateOrUpdate()),
 		).
 		Complete(r)
-}
-
-func enqueueRequestForOwnerFuncs(log logr.Logger) handler.EventHandler {
-	return &handler.Funcs{
-		CreateFunc: func(evt event.CreateEvent, q workqueue.RateLimitingInterface) {
-			enqueueRequestsForOwners(log, evt.Object, q)
-		},
-		DeleteFunc: func(evt event.DeleteEvent, q workqueue.RateLimitingInterface) {
-			enqueueRequestsForOwners(log, evt.Object, q)
-		},
-		UpdateFunc: func(e event.UpdateEvent, q workqueue.RateLimitingInterface) {
-			oldOwners := getOwnersFromObject(e.ObjectOld)
-			newOwners := getOwnersFromObject(e.ObjectNew)
-
-			for _, newOwner := range newOwners {
-				if ownerSliceContains(oldOwners, newOwner) {
-					continue
-				}
-
-				req := reconcile.Request{
-					NamespacedName: types.NamespacedName{
-						Namespace: "kyma-system",
-						Name:      newOwner.Name,
-					},
-				}
-				q.Add(req)
-				log.V(1).Info("Enqueued reconcile request for owner", "owner", req.NamespacedName)
-			}
-		},
-	}
-}
-
-func enqueueRequestsForOwners(log logr.Logger, obj metav1.Object, q workqueue.RateLimitingInterface) {
-	owners := getOwnersFromObject(obj)
-
-	for _, owner := range owners {
-		req := reconcile.Request{
-			NamespacedName: types.NamespacedName{
-				Namespace: "kyma-system",
-				Name:      owner.Name,
-			},
-		}
-		q.Add(req)
-		log.V(1).Info("Enqueued reconcile request for owner", "owner", req.NamespacedName)
-	}
-}
-
-func getOwnersFromObject(obj metav1.Object) []metav1.OwnerReference {
-	if obj == nil {
-		return nil
-	}
-
-	return obj.GetOwnerReferences()
-}
-
-func ownerSliceContains(owners []metav1.OwnerReference, owner metav1.OwnerReference) bool {
-	for _, o := range owners {
-		if o.UID == owner.UID {
-			return true
-		}
-	}
-	return false
 }
 
 func (r *LogPipelineReconciler) mapSecret(object client.Object) []reconcile.Request {
